fix(boot): remove software boot and format code
Initially, we were setting pin H3 to the blue LED, but that is the boot pin. On STM32, you can set the microcontroller to boot from flash using software, rather than hardware (pin H3 to ground), but you need to mess with the flash pins.

Some attempts were tried, but I ended up locking the flash. ST-Link Utility was used to set the STM32 in test mode 0 rather than 1 and unblock it. As a result, the hardware changes were made (pin h3 ground) and the firmware was reverted.